requirements.txt
flask==2.3.3
flask-bootstrap==3.3.7.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
concrete-ml==1.1.0